move scorecard alpha to scorecard
