--- cmd/operator-sdk/alpha/cmd.go
@@ -1,33 +1,0 @@
-// Copyright 2019 The Operator-SDK Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package alpha
-
-import (
-	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha/scorecard"
-
-	"github.com/spf13/cobra"
-)
-
-func NewCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "alpha",
-		Short: "Run an alpha subcommand",
-	}
-
-	cmd.AddCommand(
-		scorecard.NewCmd(),
-	)
-	return cmd
-}
