+++ cmd/operator-sdk/cli/legacy.go
@@ -20,13 +20,13 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/add"
-	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/alpha"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/build"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/bundle"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/completion"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/generate"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/new"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/olm"
+	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/scorecard"
 	"github.com/operator-framework/operator-sdk/cmd/operator-sdk/version"
 	"github.com/operator-framework/operator-sdk/internal/flags"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
@@ -64,9 +64,9 @@ func GetCLIRoot() *cobra.Command {
 
 	root.AddCommand(
 		add.NewCmd(),
-		alpha.NewCmd(),
 		build.NewCmd(),
 		bundle.NewCmdLegacy(),
+		scorecard.NewCmd(),
 		completion.NewCmd(),
 		generate.NewCmdLegacy(),
 		new.NewCmd(),
