+++ internal/scorecard/examples/custom-scorecard-tests/bundle/manifests/cache.example.com_memcacheds_crd.yaml
