+++ internal/scorecard/examples/custom-scorecard-tests/bundle/metadata/annotations.yaml
