+++ internal/scorecard/examples/custom-scorecard-tests/go.mod
