+++ internal/scorecard/examples/custom-scorecard-tests/images/custom-scorecard-tests/bin/entrypoint
