+++ internal/scorecard/examples/custom-scorecard-tests/images/custom-scorecard-tests/bin/user_setup
