+++ internal/scorecard/testdata/bundle.tar.gz
