+++ internal/scorecard/testdata/bundle/manifests/cache.example.com_memcacheds_crd.yaml
