+++ internal/scorecard/testdata/bundle/manifests/memcached-operator.clusterserviceversion.yaml
