+++ internal/scorecard/testdata/bundle/metadata/annotations.yaml
