+++ internal/scorecard/testdata/bundle/tests/scorecard/kuttl/kuttl-test.yaml
