+++ internal/scorecard/testdata/bundle/tests/scorecard/kuttl/list-pods/00-pod.yaml
